Fixes storing vs loading the visibility settings
When league director first connects it will attempt to set the visibility settings to the saved value. If they change in game after connection however then we will update LD to reflect that.

Also removes a bunch of extra calls to the updated signal, we were too aggresively calling this signal.